Refactor Dataset data container
The 'Data' container role has changed. It still collects data from
profilers but does not act as a storage for the dataset anymore. It is
effective for multithreaded (good encapsulation) but falls short for
manipulations (slow speed) and serialization (complex).

A new 'DataCollection' container was added to store the data in the
dataset. The metrics are now stored in a sequence which speeds up
manipulations and simplifies serialization.