rework collision_check config
visualize does not make sense in the output file.
roslaunch has to know about it as well to start rviz together with the benchmark.

output_file do not make sense in the config file.
This is specified *for a run* and not for a config.
It's the callers responsibility to put it in a sane location.